Modify mypage menu
